Switch uses of `shape.DebugString()` to `shape.ToString()` or `shape.ToProto().DebugString()`.
This prepares for removing `Shape::DebugString()` to reduce the `Shape` API surface. We cannot remove this in the same change as the function is used in the TensorFlow repo, which cannot be updated atomically with the XLA repo.

(It's unclear what format `shape.DebugString()` is in. `shape.ToProto().DebugString()` is much clearer. The new pattern also allows the caller to cache the proto for reuse.)

PiperOrigin-RevId: 753682647