add registration mechanism for all three plugins
we previously only used the registration pattern for PJRT plugins for the TPU plugin. Let's enable this for the CPU and GPU plugins, so that we can consume each of these three plugins easily internally. plan is to shift usages of the static initializer that is defined in tfrc (I don't know why there) into this shared directory.

PiperOrigin-RevId: 755988564